web_dashboard/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==5.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
PyYAML==6.0.1
colorama==0.4.6
aiosqlite==0.19.0
